fix: lint warning fast refresh
