+++ src/components/Home/Shop/Shop.jsx
@@ -1,5 +1,5 @@
 import { useReducer } from 'react'
-import { useItems } from '../../Provider/ItemsProvider'
+import { useItems } from '../../Provider/Items/Hook'
 import Card from './components/Card';
 import FilterShop from './components/FilterShop';
 import ScrollToTop from '../../ScrollToTop';
