+++ src/components/Home/Shop/components/Card.jsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import { PropTypes } from 'prop-types';
-import { useAddItemCart } from '../../../Provider/CartProvider';
+import { useAddItemCart } from '../../../Provider/Cart/Hook';
 
 function Card({ item }) {
     const onClickHandler = useAddItemCart()
